organize test cases and create a suite
