+++ Object Repository/API/backWebServices/VirtoCommerce.Platform/UserPassword/.meta
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<FolderEntity>
+   <description></description>
+   <name>UserPassword</name>
+   <tag></tag>
+   <folderType>WEBELEMENT</folderType>
+</FolderEntity>
