--- Object Repository/API/backWebServices/VirtoCommerce.Platform/_Password/.meta
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<FolderEntity>
-   <description></description>
-   <name>_Password</name>
-   <tag></tag>
-   <folderType>WEBELEMENT</folderType>
-</FolderEntity>
