--- Object Repository/API/backWebServices/VirtoCommerce.Platform/_Password/PasswordRequestResetEmail.rs
@@ -1,67 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<WebServiceRequestEntity>
-   <description></description>
-   <name>PasswordRequestResetEmail</name>
-   <tag></tag>
-   <elementGuidId>4cd642ae-f325-4889-a9f2-6fafc1e9ca69</elementGuidId>
-   <selectorMethod>BASIC</selectorMethod>
-   <useRalativeImagePath>false</useRalativeImagePath>
-   <connectionTimeout>-1</connectionTimeout>
-   <followRedirects>false</followRedirects>
-   <httpBody></httpBody>
-   <httpBodyContent>{
-  &quot;text&quot;: &quot;//body is empty in swagger&quot;,
-  &quot;contentType&quot;: &quot;application/json&quot;,
-  &quot;charset&quot;: &quot;UTF-8&quot;
-}</httpBodyContent>
-   <httpBodyType>text</httpBodyType>
-   <httpHeaderProperties>
-      <isSelected>true</isSelected>
-      <matchCondition>equals</matchCondition>
-      <name>Content-Type</name>
-      <type>Main</type>
-      <value>application/json</value>
-   </httpHeaderProperties>
-   <httpHeaderProperties>
-      <isSelected>true</isSelected>
-      <matchCondition>equals</matchCondition>
-      <name>api_key</name>
-      <type>Main</type>
-      <value>${GlobalVariable.api_key}</value>
-   </httpHeaderProperties>
-   <maxResponseSize>-1</maxResponseSize>
-   <migratedVersion>5.4.1</migratedVersion>
-   <restRequestMethod>POST</restRequestMethod>
-   <restUrl>${GlobalVariable.urlBack}/api/platform/security/users/${loginOrEmail}/requestpasswordreset</restUrl>
-   <serviceType>RESTful</serviceType>
-   <soapBody></soapBody>
-   <soapHeader></soapHeader>
-   <soapRequestMethod></soapRequestMethod>
-   <soapServiceEndpoint></soapServiceEndpoint>
-   <soapServiceFunction></soapServiceFunction>
-   <socketTimeout>-1</socketTimeout>
-   <useServiceInfoFromWsdl>true</useServiceInfoFromWsdl>
-   <variables>
-      <defaultValue>'QweUser'</defaultValue>
-      <description></description>
-      <id>3bc88c64-ed36-47f6-b0ce-59103054f46e</id>
-      <masked>false</masked>
-      <name>loginOrEmail</name>
-   </variables>
-   <verificationScript>import static org.assertj.core.api.Assertions.*
-
-import com.kms.katalon.core.testobject.RequestObject
-import com.kms.katalon.core.testobject.ResponseObject
-import com.kms.katalon.core.webservice.keyword.WSBuiltInKeywords as WS
-import com.kms.katalon.core.webservice.verification.WSResponseManager
-
-import groovy.json.JsonOutput
-import com.kms.katalon.core.util.KeywordUtil
-
-
-ResponseObject response = WSResponseManager.getInstance().getCurrentResponse()
-def prettyJson = JsonOutput.prettyPrint(response.getResponseBodyContent())
-KeywordUtil.logInfo(prettyJson)
-WS.verifyResponseStatusCode(response, 200)</verificationScript>
-   <wsdlAddress></wsdlAddress>
-</WebServiceRequestEntity>
