--- Object Repository/API/backWebServices/VirtoCommerce.Platform/_Password/PasswordValidate.rs
@@ -1,66 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<WebServiceRequestEntity>
-   <description></description>
-   <name>PasswordValidate</name>
-   <tag></tag>
-   <elementGuidId>79bbb741-c6e2-4d8f-b98a-33269c5a4178</elementGuidId>
-   <selectorMethod>BASIC</selectorMethod>
-   <useRalativeImagePath>false</useRalativeImagePath>
-   <connectionTimeout>-1</connectionTimeout>
-   <followRedirects>false</followRedirects>
-   <httpBody></httpBody>
-   <httpBodyContent>{
-  &quot;text&quot;: &quot;\&quot;${password}\&quot;\n&quot;,
-  &quot;contentType&quot;: &quot;application/json&quot;,
-  &quot;charset&quot;: &quot;UTF-8&quot;
-}</httpBodyContent>
-   <httpBodyType>text</httpBodyType>
-   <httpHeaderProperties>
-      <isSelected>true</isSelected>
-      <matchCondition>equals</matchCondition>
-      <name>Content-Type</name>
-      <type>Main</type>
-      <value>application/json</value>
-   </httpHeaderProperties>
-   <httpHeaderProperties>
-      <isSelected>true</isSelected>
-      <matchCondition>equals</matchCondition>
-      <name>Accept</name>
-      <type>Main</type>
-      <value>text/plain</value>
-   </httpHeaderProperties>
-   <maxResponseSize>-1</maxResponseSize>
-   <migratedVersion>5.4.1</migratedVersion>
-   <restRequestMethod>POST</restRequestMethod>
-   <restUrl>${GlobalVariable.urlBack}/api/platform/security/validatepassword</restUrl>
-   <serviceType>RESTful</serviceType>
-   <soapBody></soapBody>
-   <soapHeader></soapHeader>
-   <soapRequestMethod></soapRequestMethod>
-   <soapServiceEndpoint></soapServiceEndpoint>
-   <soapServiceFunction></soapServiceFunction>
-   <socketTimeout>-1</socketTimeout>
-   <useServiceInfoFromWsdl>true</useServiceInfoFromWsdl>
-   <variables>
-      <defaultValue>''</defaultValue>
-      <description></description>
-      <id>538696fd-ed68-4e76-adde-1540b721b0c4</id>
-      <masked>false</masked>
-      <name>password</name>
-   </variables>
-   <verificationScript>import static org.assertj.core.api.Assertions.*
-
-import com.kms.katalon.core.testobject.ResponseObject
-import com.kms.katalon.core.webservice.keyword.WSBuiltInKeywords as WS
-import com.kms.katalon.core.webservice.verification.WSResponseManager
-
-import groovy.json.JsonOutput
-import com.kms.katalon.core.util.KeywordUtil
-
-
-ResponseObject response = WSResponseManager.getInstance().getCurrentResponse()
-def prettyJson = JsonOutput.prettyPrint(response.getResponseBodyContent())
-KeywordUtil.logInfo(prettyJson)
-WS.verifyResponseStatusCode(response, 200)</verificationScript>
-   <wsdlAddress></wsdlAddress>
-</WebServiceRequestEntity>
