+++ Scripts/API Coverage/Contacts/_module_Customer_Member/MemberCreationInOrgWithAccount/Script1634663763212.groovy
@@ -79,7 +79,7 @@ WS.sendRequestAndVerify(findTestObject('API/backWebServices/VirtoCommerce.Platfo
 
 
 WebUI.comment("TEST CASE : Update user password as admin way")
-responseResetPassword = WS.sendRequestAndVerify(findTestObject('API/backWebServices/VirtoCommerce.Platform/ResetPassword', [
+responseResetPassword = WS.sendRequestAndVerify(findTestObject('API/backWebServices/VirtoCommerce.Platform/UserPassword/UserPasswordReset', [
 	('userName') : GlobalVariable.userName
 	]))
 WS.verifyElementPropertyValue(responseResetPassword, 'succeeded', true)
