--- Scripts/API Coverage/Platform/Password/PasswordResetOnLoginPage/Script1657866043761.groovy
@@ -1,56 +1,0 @@
-import static com.kms.katalon.core.checkpoint.CheckpointFactory.findCheckpoint
-import static com.kms.katalon.core.testcase.TestCaseFactory.findTestCase
-import static com.kms.katalon.core.testdata.TestDataFactory.findTestData
-import static com.kms.katalon.core.testobject.ObjectRepository.findTestObject
-import static com.kms.katalon.core.testobject.ObjectRepository.findWindowsObject
-import com.kms.katalon.core.checkpoint.Checkpoint as Checkpoint
-import com.kms.katalon.core.cucumber.keyword.CucumberBuiltinKeywords as CucumberKW
-import com.kms.katalon.core.mobile.keyword.MobileBuiltInKeywords as Mobile
-import com.kms.katalon.core.model.FailureHandling as FailureHandling
-import com.kms.katalon.core.testcase.TestCase as TestCase
-import com.kms.katalon.core.testdata.TestData as TestData
-import com.kms.katalon.core.testng.keyword.TestNGBuiltinKeywords as TestNGKW
-import com.kms.katalon.core.testobject.TestObject as TestObject
-import com.kms.katalon.core.webservice.keyword.WSBuiltInKeywords as WS
-import com.kms.katalon.core.webui.keyword.WebUiBuiltInKeywords as WebUI
-import com.kms.katalon.core.windows.keyword.WindowsBuiltinKeywords as Windows
-import internal.GlobalVariable as GlobalVariable
-import org.openqa.selenium.Keys as Keys
-
-
-import groovy.json.JsonBuilder
-import groovy.json.JsonSlurper
-import com.kms.katalon.core.testobject.impl.HttpTextBodyContent
-import com.kms.katalon.core.testobject.RequestObject
-import groovy.json.JsonOutput
-import com.kms.katalon.core.testobject.ResponseObject
-
-
-WebUI.comment('TEST CASE: Change or reset password via request on login page')
-//this operations can not be fully tested due to Katalon <> Platform integration limitations
-
-
-'REQUEST PASSWORD RESET BY LOGIN OR EMAIL'//simple check for 200ok
-passwordRequestReset = WS.sendRequestAndVerify(findTestObject('API/backWebServices/VirtoCommerce.Platform/_Password/PasswordRequestResetEmail'))
-//WS.verifyElementPropertyValue(passwordRequestReset, 'succeeded', true)
-
-
-'VALIDATE PASSWORD RESET TOKEN'//only checks for 200ok for now
-passwordValidateResetToken = WS.sendRequestAndVerify(findTestObject('Object Repository/API/backWebServices/VirtoCommerce.Platform/_Password/PasswordValidateResetToken',[
-	('userId') : GlobalVariable.userId,
-	('token') : GlobalVariable.token
-	]))
-//WS.verifyElementPropertyValue(response, 'succeeded', true)
-
-
-'CONFIRM RESET PASSWORD'//only checks for 200ok for now
-newPassword = GlobalVariable.userPassword + '!'
-passwordResetConfirm = WS.sendRequestAndVerify(findTestObject('API/backWebServices/VirtoCommerce.Platform/_Password/PasswordResetConfirm',[
-	('newPassword') : newPassword,
-	('userId') : GlobalVariable.userId,
-	('token') : GlobalVariable.token
-	]))
-//WS.verifyElementPropertyValue(response, 'succeeded', true)
-
-
-
