--- Test Cases/API Coverage/Platform/Password/PasswordOperationsAdminPath.tc
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<TestCaseEntity>
-   <description></description>
-   <name>PasswordOperationsAdminPath</name>
-   <tag></tag>
-   <comment></comment>
-   <testCaseGuid>4733fd83-9dd4-4cb4-909d-d7b9a6c2739a</testCaseGuid>
-</TestCaseEntity>
