--- Test Cases/API Coverage/Platform/Password/PasswordOperationsCurrentUser.tc
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<TestCaseEntity>
-   <description></description>
-   <name>PasswordOperationsCurrentUser</name>
-   <tag></tag>
-   <comment></comment>
-   <testCaseGuid>43ec9d61-6424-4567-be65-4c35bda90b34</testCaseGuid>
-</TestCaseEntity>
