--- Test Cases/API Coverage/Platform/Password/PasswordResetOnLoginPage.tc
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<TestCaseEntity>
-   <description></description>
-   <name>PasswordResetOnLoginPage</name>
-   <tag></tag>
-   <comment></comment>
-   <testCaseGuid>76d4c35b-3c68-47cd-aaf4-0c1fc7108946</testCaseGuid>
-</TestCaseEntity>
