--- Test Cases/API Coverage/Platform/Password/PasswordValidate(2ReplaceScript).tc
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<TestCaseEntity>
-   <description>Validate password / check returned errors</description>
-   <name>PasswordValidate(2ReplaceScript)</name>
-   <tag></tag>
-   <comment></comment>
-   <testCaseGuid>3d8f7c90-6433-415c-94af-9b0d54f4cd97</testCaseGuid>
-</TestCaseEntity>
