--- Test Cases/API Coverage/Platform/Password/PasswordValidate.tc
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<TestCaseEntity>
-   <description>Validate password / check returned errors</description>
-   <name>PasswordValidate</name>
-   <tag></tag>
-   <comment></comment>
-   <testCaseGuid>cfb6c535-e470-4064-bfa7-7d25907b10a0</testCaseGuid>
-</TestCaseEntity>
