+++ Test Cases/API Coverage/Platform/UserPassword/.meta
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<FolderEntity>
+   <description></description>
+   <name>UserPassword</name>
+   <tag></tag>
+   <folderType>TESTCASE</folderType>
+</FolderEntity>
