+++ Test Cases/API Coverage/Platform/UserPassword/UserPasswordOperationsAdminPath.tc
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<TestCaseEntity>
+   <description>Password operations admin > other users</description>
+   <name>UserPasswordOperationsAdminPath</name>
+   <tag></tag>
+   <comment></comment>
+   <testCaseGuid>4733fd83-9dd4-4cb4-909d-d7b9a6c2739a</testCaseGuid>
+</TestCaseEntity>
