+++ Test Cases/API Coverage/Platform/UserPassword/UserPasswordOperationsCurrentUser.tc
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<TestCaseEntity>
+   <description>Passwod operations as a logged user</description>
+   <name>UserPasswordOperationsCurrentUser</name>
+   <tag></tag>
+   <comment></comment>
+   <testCaseGuid>43ec9d61-6424-4567-be65-4c35bda90b34</testCaseGuid>
+</TestCaseEntity>
