+++ Test Cases/API Coverage/Platform/UserPassword/UserPasswordResetOnLoginPage.tc
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<TestCaseEntity>
+   <description>Enpoints used when resetting password from login page (forgot password functionality)</description>
+   <name>UserPasswordResetOnLoginPage</name>
+   <tag></tag>
+   <comment></comment>
+   <testCaseGuid>76d4c35b-3c68-47cd-aaf4-0c1fc7108946</testCaseGuid>
+</TestCaseEntity>
