+++ Test Cases/API Coverage/Platform/UserPassword/_suiteTemplate.tc
