+++ Test Suites/Modules/_DRAFT_module_Platform_User_Management.ts
@@ -1,0 +1,80 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<TestSuiteEntity>
+   <description></description>
+   <name>_DRAFT_module_Platform_User_Management</name>
+   <tag></tag>
+   <isRerun>false</isRerun>
+   <mailRecipient></mailRecipient>
+   <numberOfRerun>0</numberOfRerun>
+   <pageLoadTimeout>10</pageLoadTimeout>
+   <pageLoadTimeoutDefault>true</pageLoadTimeoutDefault>
+   <rerunFailedTestCasesOnly>false</rerunFailedTestCasesOnly>
+   <rerunImmediately>false</rerunImmediately>
+   <testSuiteGuid>e73ca6e4-8689-4522-8b6d-4f653dc8f050</testSuiteGuid>
+   <testCaseLink>
+      <guid>373d7656-3acb-4352-8eef-228ed074e2ee</guid>
+      <isReuseDriver>false</isReuseDriver>
+      <isRun>true</isRun>
+      <testCaseId>Test Cases/API Coverage/Platform/AuthBackToken</testCaseId>
+   </testCaseLink>
+   <testCaseLink>
+      <guid>29b89881-58ea-42a8-9b1a-61cc19f392d8</guid>
+      <isReuseDriver>false</isReuseDriver>
+      <isRun>true</isRun>
+      <testCaseId>Test Cases/API Coverage/Platform/ApiKeySet</testCaseId>
+   </testCaseLink>
+   <testCaseLink>
+      <guid>efe8795f-1f8d-4796-b45b-0870c8c3b83f</guid>
+      <isReuseDriver>false</isReuseDriver>
+      <isRun>true</isRun>
+      <testCaseId>Test Cases/API Coverage/Platform/UserGetAllRequestsCheck</testCaseId>
+   </testCaseLink>
+   <testCaseLink>
+      <guid>7f7ae5fb-6440-4099-a559-98ac7129561c</guid>
+      <isReuseDriver>false</isReuseDriver>
+      <isRun>true</isRun>
+      <testCaseId>Test Cases/API Coverage/Platform/UserCreateAdminAndGetToken</testCaseId>
+   </testCaseLink>
+   <testCaseLink>
+      <guid>c38d7c1f-b2d2-4548-b962-710cc5db8087</guid>
+      <isReuseDriver>false</isReuseDriver>
+      <isRun>true</isRun>
+      <testCaseId>Test Cases/API Coverage/Platform/UserCreateAlreadyExisted</testCaseId>
+   </testCaseLink>
+   <testCaseLink>
+      <guid>1ccea60e-1b02-4e23-ba4d-bae9a4fa1ccd</guid>
+      <isReuseDriver>false</isReuseDriver>
+      <isRun>true</isRun>
+      <testCaseId>Test Cases/API Coverage/Platform/UserSendVerifictaionEmail</testCaseId>
+   </testCaseLink>
+   <testCaseLink>
+      <guid>3e81f9a6-299c-4e87-b35a-6b702ec3b85c</guid>
+      <isReuseDriver>false</isReuseDriver>
+      <isRun>true</isRun>
+      <testCaseId>Test Cases/API Coverage/Platform/UserPassword/UserPasswordValidate</testCaseId>
+   </testCaseLink>
+   <testCaseLink>
+      <guid>15a2c547-b243-4573-afaf-ec65882ea082</guid>
+      <isReuseDriver>false</isReuseDriver>
+      <isRun>true</isRun>
+      <testCaseId>Test Cases/API Coverage/Platform/UserPassword/UserPasswordResetOnLoginPage</testCaseId>
+   </testCaseLink>
+   <testCaseLink>
+      <guid>bbc3d46a-fb27-4b0f-9925-e9f3ac083ff6</guid>
+      <isReuseDriver>false</isReuseDriver>
+      <isRun>true</isRun>
+      <testCaseId>Test Cases/API Coverage/Platform/UserPassword/UserPasswordOperationsAdminPath</testCaseId>
+   </testCaseLink>
+   <testCaseLink>
+      <guid>0b3bc822-14ba-407e-b8b3-94a600464a6f</guid>
+      <isReuseDriver>false</isReuseDriver>
+      <isRun>true</isRun>
+      <testCaseId>Test Cases/API Coverage/Platform/UserPassword/UserPasswordOperationsCurrentUser</testCaseId>
+   </testCaseLink>
+   <testCaseLink>
+      <guid>9bcba098-3799-4129-aae2-6e180a6c5aaa</guid>
+      <isReuseDriver>false</isReuseDriver>
+      <isRun>true</isRun>
+      <testCaseId>Test Cases/API Coverage/Platform/UserDelete</testCaseId>
+   </testCaseLink>
+</TestSuiteEntity>
